move example to new file
